refactor[leaderboard]: used a mixin to combine murderboard/nuggieboard code
